+++ src/components/EventLoop/EventLoop.js
@@ -1,9 +1,5 @@
 import React, { Component } from 'react'
 import './EventLoop.styles.css'
-import { connect } from 'react-redux'
-
-import { removeFromCallbackQueue } from '../../redux/callbackQueue/callbackQueue.actions'
-import { addFunctionToCallstack } from '../../redux/callstack/callstack.actions'
 
 class EventLoop extends Component {
 	constructor(props) {
@@ -11,32 +7,6 @@ class EventLoop extends Component {
 		this.state = {
 			spin: false,
 		}
-
-		this.makeFuncObj = (str) => ({
-			name: str,
-			delay: 0,
-			webApi: false,
-			message: 'Test Console log',
-		})
-
-		this.runLoop = () => {
-			// if callstack is empty and callbackqueue is not empty
-			if (this.props.callstack.length === 0 && this.props.callbackQueue.length !== 0) {
-				// add the callbackqueue[0] to callstack
-				// but callbackqueue has func as strings whereas
-				// callstack accepts func as obj (see callstack reducer)
-				// hence wrap the string in makeFuncObj
-				this.props.addFunctionToCallstack(
-					this.makeFuncObj(this.props.callbackQueue[0] + 'added by event loop')
-				)
-				this.props.removeFromCallbackQueue()
-			}
-		}
-
-		this.timerId = setInterval(() => {
-			this.runLoop()
-			console.log('ran loop')
-		}, 1000)
 	}
 
 	render() {
@@ -55,14 +25,4 @@ class EventLoop extends Component {
 	}
 }
 
-const mapStateToProps = (state) => ({
-	callstack: state.callstack.stack,
-	callbackQueue: state.callbackQueue,
-})
-
-const mapDispatchToProps = (dispatch) => ({
-	removeFromCallbackQueue: () => dispatch(removeFromCallbackQueue()),
-	addFunctionToCallstack: (func) => dispatch(addFunctionToCallstack(func)),
-})
-
-export default connect(mapStateToProps, mapDispatchToProps)(EventLoop)
+export default EventLoop
